+++ controllers/management_cluster_network_test.go
@@ -12,7 +12,6 @@ import (
 	. "github.com/onsi/gomega/gstruct"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/types"
 	k8stypes "k8s.io/apimachinery/pkg/types"
 	capa "sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta2"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
@@ -268,7 +267,7 @@ var _ = Describe("ManagementClusterTransitGatewayReconciler", func() {
 			})
 
 			It("removes the finalizer", func() {
-				err := k8sClient.Get(context.Background(), types.NamespacedName{Name: cluster.Name, Namespace: namespace}, cluster)
+				err := k8sClient.Get(context.Background(), k8stypes.NamespacedName{Name: cluster.Name, Namespace: namespace}, cluster)
 				Expect(k8serrors.IsNotFound(err)).To(BeTrue())
 			})