+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.0] - 2025-06-16
+
 ### Added
 
 - Add controller to create node pool bootstrap data on S3.
@@ -249,7 +251,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - changed: `app.giantswarm.io` label group was changed to `application.giantswarm.io`
 
-[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.17.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.14.6...v0.15.0